Renamed the UserSearchRequest file
